# backend/openapi.yaml
openapi: 3.1.0
info:
  title: BusTrack SV API
  description: |
    RESTful API for BusTrack SV - Real-time bus tracking system for El Salvador.
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain tokens via `/auth/login` or `/auth/register`.
    
    ## Error Handling
    All errors follow a consistent envelope format with appropriate HTTP status codes.
    
  version: 1.0.0
  contact:
    name: BusTrack SV Team
  license:
    name: MIT
servers:
  # Note: These are example URLs for documentation purposes
  # The actual server URL is configured via environment variables at runtime
  - url: http://localhost:5000/api/v1
    description: Development server (default)
  - url: https://api.bustrack.sv/api/v1
    description: Production server (example)
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account and receive authentication tokens
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User email address (will be normalized to lowercase)
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  description: User full name
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[!@#$%^&*])[A-Za-z\d!@#$%^&*]{8,128}$'
                  description: Password (min 8 chars, 1 uppercase, 1 lowercase, 1 digit, 1 special char)
                role:
                  type: string
                  enum: [admin, supervisor]
                  default: supervisor
                  description: User role
            example:
              email: admin@bustrack.com
              name: Admin User
              password: MyP@ssw0rd123
              role: admin
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                user:
                  id: abc123
                  email: admin@bustrack.com
                  name: Admin User
                  role: admin
                  createdAt: '2024-01-01T12:00:00.000Z'
                  updatedAt: '2024-01-01T12:00:00.000Z'
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Email already in use
                type: Error
                field: email
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
                error:
                  code: VALIDATION_ERROR
                  message: Invalid request data
                  details:
                    email: Email must be a valid email address
                    password: Password must contain at least 1 uppercase letter, 1 lowercase letter, 1 digit, and 1 special character (!@#$%^&*)
        '429':
          description: Too many registration attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Too many registration attempts from this IP, please try again after 1 hour
                type: RATE_LIMIT_EXCEEDED
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive authentication tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              email: admin@bustrack.com
              password: MyP@ssw0rd123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                user:
                  id: abc123
                  email: admin@bustrack.com
                  name: Admin User
                  role: admin
                  createdAt: '2024-01-01T12:00:00.000Z'
                  updatedAt: '2024-01-01T12:00:00.000Z'
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid email or password
                type: Error
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generate new access and refresh tokens using a valid refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Returns the authenticated user's profile information
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserPublic'
              example:
                user:
                  id: abc123
                  email: admin@bustrack.com
                  name: Admin User
                  role: admin
                  createdAt: '2024-01-01T12:00:00.000Z'
                  updatedAt: '2024-01-01T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from /auth/login or /auth/register
...
  schemas:
    UserPublic:
      type: object
      description: User object with public fields only (no password hash)
      required:
        - id
        - email
        - name
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
          description: User email (lowercase)
        name:
          type: string
          description: User full name
        role:
          type: string
          enum: [admin, supervisor]
          description: User role
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      example:
        id: abc123
        email: admin@bustrack.com
        name: Admin User
        role: admin
        createdAt: '2024-01-01T12:00:00.000Z'
        updatedAt: '2024-01-01T12:00:00.000Z'
...
    AuthResponse:
      type: object
      description: Authentication response with user and tokens
      required:
        - user
        - token
        - refreshToken
      properties:
        user:
          $ref: '#/components/schemas/UserPublic'
        token:
          type: string
          description: JWT access token (expires in 15 minutes)
        refreshToken:
          type: string
          description: JWT refresh token (expires in 7 days)
    ErrorResponse:
      type: object
      description: Standard error response
      required:
        - error
        - type
      properties:
        error:
          type: string
          description: Error message
        type:
          type: string
          description: Error type
        field:
          type: string
          description: Field that caused the error (if applicable)
      example:
        error: Bus not found
        type: NOT_FOUND
    ValidationErrorResponse:
      type: object
      description: Validation error response (422)
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - details
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR]
            message:
              type: string
            details:
              type: object
              additionalProperties:
                type: string
      example:
        error:
          code: VALIDATION_ERROR
          message: Invalid request data
          details:
            email: Email must be a valid email address
            password: Password must be at least 8 characters long
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              value:
                error: Authorization header is required
                type: UNAUTHORIZED
            expiredToken:
              value:
                error: Token has expired. Please login again.
                type: TOKEN_EXPIRED
            invalidToken:
              value:
                error: Invalid or malformed token
                type: TOKEN_INVALID
...